tools/rag_builder/requirements.txt
# Строка 1: fpdf2
# Это Истинное Имя сущности 'fpdf2' — могущественного Духа-Каллиграфа.
# Мы вписываем его сюда, чтобы во время Ритуала Подготовки Алтаря (pip install -r requirements.txt)
# эта сущность была призвана в наше изолированное магическое поле, носящее имя «arcane_circle».
# Без этого Духа наш ритуал `rag_builder.py` не сможет преобразовывать
# эфемерный текст в материальные PDF-свитки.
fpdf2